refactor(image): defer scim-plugin scopes creation when running config-api
