+++ docker-jans-config-api/scripts/utils.py
@@ -1,3 +1,5 @@
+import os
+
 import ruamel.yaml
 
 
@@ -6,3 +8,7 @@ def parse_config_api_swagger(path="/app/static/jans-config-api-swagger-auto.yaml
         txt = f.read()
     txt = txt.replace("\t", " ")
     return ruamel.yaml.load(txt, Loader=ruamel.yaml.RoundTripLoader)
+
+
+def generate_hex(size: int = 3):
+    return os.urandom(size).hex().upper()
